add data rest test
